move utils folder into lib
