+++ src/app/(app)/cookbook/page.tsx
@@ -2,17 +2,17 @@
 
 import CookbookChapter from '@/app/ui/CookbookChapter'
 import React, { useContext, useEffect, useState } from 'react'
-import { SessionContext, updateCookbook } from '@/app/utils/Session'
+import { SessionContext, updateCookbook } from '@/app/lib/utils/Session'
 import { NewRecipe } from '@/app/lib/types'
 import ThemedButton from '@/app/ui/buttons/ThemedButton'
 import AddIcon from '@mui/icons-material/Add'
 import DeleteChapterDialog from '@/app/ui/dialogs/DeleteChapterDialog'
 import EditRecipeDialog from '@/app/ui/dialogs/EditRecipeDialog'
-import { addNewChapter } from '@/app/utils/addNewChapter'
-import { addNewRecipe } from '@/app/utils/addNewRecipe'
-import { deleteChapter } from '@/app/utils/deleteChapter'
+import { addNewChapter } from '@/app/lib/utils/dbHelpers/addNewChapter'
+import { addNewRecipe } from '@/app/lib/utils/dbHelpers/addNewRecipe'
+import { deleteChapter } from '@/app/lib/utils/dbHelpers/deleteChapter'
 import { DragDropContext, Draggable, Droppable, DropResult } from '@hello-pangea/dnd'
-import { users } from '@/app/utils/firebase'
+import { users } from '@/app/lib/utils/firebase'
 import clsx from 'clsx'
 import { redirect } from 'next/navigation'
 import LoadingIcon from '@/app/ui/LoadingIcon'
