+++ src/app/ui/CookbookChapter.tsx
@@ -4,13 +4,13 @@ import AccordionDetails from '@mui/material/AccordionDetails'
 import ExpandCircleDownIcon from '@mui/icons-material/ExpandCircleDown'
 import CookbookRecipe from './CookbookRecipe'
 import { RecipeWithNotes } from '../lib/types'
-import { users } from '../utils/firebase'
+import { users } from '../lib/utils/firebase'
 import { useContext, MouseEvent } from 'react'
-import { SessionContext } from '../utils/Session'
+import { SessionContext } from '../lib/utils/Session'
 import InlineInput from './inputs/InlineInput'
 import { arrayRemove } from 'firebase/firestore'
 import DeleteRoundedIcon from '@mui/icons-material/DeleteRounded'
-import { sharedMiniButtonStyles } from '../utils/sharedMiniButtonStyles'
+import { sharedMiniButtonStyles } from '../lib/styles/sharedMiniButtonStyles'
 import { CookbookChapterProps } from '../lib/types/ui'
 import { Droppable, Draggable } from '@hello-pangea/dnd'
 import clsx from 'clsx'
