+++ src/app/ui/CookbookRecipe.tsx
@@ -3,8 +3,8 @@ import Accordion from '@mui/material/Accordion'
 import AccordionSummary from '@mui/material/AccordionSummary'
 import { useContext } from 'react'
 import InlineInput from './inputs/InlineInput'
-import { users } from '../utils/firebase'
-import { SessionContext } from '../utils/Session'
+import { users } from '../lib/utils/firebase'
+import { SessionContext } from '../lib/utils/Session'
 import { arrayRemove } from 'firebase/firestore'
 import CookbookNotes from './CookbookNotes'
 import RecipeMenu from './RecipeMenu'
