+++ src/app/ui/buttons/GoogleSignInButton.tsx
@@ -1,7 +1,7 @@
 import { GoogleSignInButtonProps } from '@/app/lib/types/ui/buttons'
 import { getAdditionalUserInfo, GoogleAuthProvider, signInWithPopup } from 'firebase/auth'
-import { auth } from '../../utils/firebase'
-import { createUser } from '@/app/utils/createUser'
+import { auth } from '../../lib/utils/firebase'
+import { createUser } from '@/app/lib/utils/dbHelpers/createUser'
 import { FirebaseUser } from '@/app/lib/types'
 import ThemedButton from './ThemedButton'
 import Image from 'next/image'
