+++ src/app/ui/dialogs/EditRecipeDialog.tsx
@@ -2,7 +2,7 @@ import { DialogActions, DialogContent, DialogTitle } from '@mui/material'
 import ThemedButton from '../buttons/ThemedButton'
 import ThemedTextField from '../inputs/ThemedTextField'
 import { ChangeEvent, useContext, useEffect, useState } from 'react'
-import { SessionContext } from '@/app/utils/Session'
+import { SessionContext } from '@/app/lib/utils/Session'
 import { EditRecipeDialogProps } from '@/app/lib/types/ui/dialogs'
 import BaseDialog from './BaseDialog'
