+++ src/app/ui/dialogs/SearchDialog.tsx
@@ -2,7 +2,7 @@ import { DialogContent, DialogTitle, InputAdornment } from '@mui/material'
 import ThemedTextField from '../inputs/ThemedTextField'
 import SearchIcon from '@mui/icons-material/Search'
 import { ChangeEvent, useContext, useEffect, useState } from 'react'
-import { SessionContext } from '@/app/utils/Session'
+import { SessionContext } from '@/app/lib/utils/Session'
 import { SearchRecipe } from '@/app/lib/types'
 import CookbookRecipe from '../CookbookRecipe'
 import { SearchDialogProps } from '@/app/lib/types/ui/dialogs'
